requirements.txt
Django>=5.2.8
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.10
django-apscheduler>=0.6.2
apscheduler>=3.10.4
gunicorn
psycopg2-binary
whitenoise
dj-database-url
psycopg2-binary